+++ src/main/java/com/backend/tienda/entidades/Productos.java
@@ -17,15 +17,15 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor @NoArgsConstructor @Builder
 public class Productos {
     @Id
-    private Double codigo_producto;
+    private Long codigo_producto;
 
-    private Double ivacompra;
+    private Long ivacompra;
 
     @ManyToOne(optional = false)
     @JoinColumn(name = "nitproveedor")
     private Proveedores nitproveedor;
 
     private String nombre_producto;
-    private Double precio_compra;
-    private Double precio_venta;
+    private Long precio_compra;
+    private Long precio_venta;
 }
