+++ src/main/java/com/backend/tienda/repository/UsuariosRepository.java
@@ -3,5 +3,5 @@ package com.backend.tienda.repository;
 import com.backend.tienda.entidades.Usuarios;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UsuariosRepository extends JpaRepository<Usuarios, Double> {
+public interface UsuariosRepository extends JpaRepository<Usuarios, Long> {
 }
